# api_gateway/swagger.yaml
openapi: 3.0.0
info:
  title: API Gateway Service
  version: 1.0.0
  description: Entry point for all notification requests, handling validation, authentication, routing to message queues, and tracking notification status.
servers:
  - url: http://localhost:3000
    description: Local Development Server
tags:
  - name: Notifications
    description: Operations related to sending and managing notifications.
  - name: Health
    description: Health check endpoint for service status.
components:
  schemas:
    NotificationRequest:
      type: object
      required:
        - user_id
        - template_code
        - notification_type
        - variables
      properties:
        user_id:
          type: string
          format: uuid
          description: The UUID of the user to send the notification to.
          example: "550e8400-e29b-41d4-a716-446655440000"
        template_code:
          type: string
          description: The template code/name to use for the notification (e.g., "welcome_email").
          example: "welcome_email"
        notification_type:
          type: string
          enum:
            - email
            - push
          description: The type of notification to send.
          example: "email"
        variables:
          type: object
          description: Variables for template substitution (UserData format).
          properties:
            name:
              type: string
              description: User's name
              example: "John Doe"
            link:
              type: string
              format: uri
              description: URL link to include in notification
              example: "https://example.com/welcome"
            meta:
              type: object
              description: Optional metadata
              additionalProperties: true
          required:
            - name
            - link
        request_id:
          type: string
          description: Optional unique request identifier for idempotency.
          example: "req-12345-67890"
        priority:
          type: integer
          description: Optional priority level (0-10, higher is more important).
          example: 1
        metadata:
          type: object
          description: Optional additional metadata.
          additionalProperties: true
          example:
            source: "api"
            campaign: "welcome_flow"
    NotificationStatusResponse:
      type: object
      required:
        - notification_id
        - user_id
        - template_code
        - notification_type
        - status
        - timestamp
        - details
      properties:
        notification_id:
          type: string
          format: uuid
          description: The unique ID of the notification request (correlation ID).
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        user_id:
          type: string
          format: uuid
          description: The UUID of the user.
          example: "550e8400-e29b-41d4-a716-446655440000"
        template_code:
          type: string
          description: The template code/name used for the notification.
          example: "welcome_email"
        notification_type:
          type: string
          enum:
            - email
            - push
          description: The type of notification.
          example: "email"
        status:
          type: string
          enum:
            - PENDING
            - QUEUED
            - SENT
            - FAILED
          description: The current status of the notification.
          example: "QUEUED"
        timestamp:
          type: string
          format: date-time
          description: The timestamp when the status was last updated.
          example: "2025-11-13T10:00:00Z"
        details:
          type: string
          description: Additional details about the notification status.
          example: "Notification message published to RabbitMQ."
    GenericSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        message:
          type: string
          example: "Operation successful."
        meta:
          type: object
          nullable: true
    GenericErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Internal Server Error"
        message:
          type: string
          example: "Failed to process request."
        meta:
          type: object
          nullable: true
paths:
  /notifications/send:
    post:
      summary: Send a new notification
      tags:
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
      responses:
        '202':
          description: Notification request accepted and queued for processing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      notification_id:
                        type: string
                        format: uuid
                        example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                      status:
                        type: string
                        example: "accepted"
                      routing_key:
                        type: string
                        example: "email"
                  message:
                    type: string
                    example: "Notification request accepted and queued for processing."
                  meta:
                    type: object
                    nullable: true
        '400':
          description: Bad Request (e.g., invalid notification_type, missing fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
  /notifications/{notification_id}/status:
    get:
      summary: Get notification status by ID
      tags:
        - Notifications
      parameters:
        - in: path
          name: notification_id
          schema:
            type: string
            format: uuid
          required: true
          description: The unique ID of the notification request.
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
      responses:
        '200':
          description: Notification status retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/NotificationStatusResponse'
        '404':
          description: Notification status not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
  /{notification_preference}/status:
    post:
      summary: Update notification status
      tags:
        - Notifications
      parameters:
        - in: path
          name: notification_preference
          schema:
            type: string
            enum: [email, push]
          required: true
          description: The type of notification preference (email or push).
          example: "email"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - notification_id
                - status
              properties:
                notification_id:
                  type: string
                  description: The unique ID of the notification request.
                  example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                status:
                  type: string
                  enum: [delivered, pending, failed]
                  description: The status of the notification.
                  example: "delivered"
                timestamp:
                  type: string
                  format: date-time
                  description: Optional timestamp when the status was updated.
                  example: "2024-01-01T12:00:00Z"
                error:
                  type: string
                  description: Optional error message if status is failed.
                  example: "Failed to send email: Invalid recipient"
      responses:
        '200':
          description: Notification status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/NotificationStatusResponse'
        '400':
          description: Bad Request (e.g., invalid status, missing fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
        '404':
          description: Notification status not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorResponse'
  /health:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: API Gateway is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"